🔥 fix(ingest/scheduler): corregir desfase UTC-4 en estaciones meteorologicas y script de sobreescritura in-situ
- [index.ts](file:///c:/Dev/pristinoplant/services/ingest/src/index.ts): Generalizar condicion de ajuste de zona horaria UTC-4 (+4h) para todas las fuentes `Weather_Station` independientemente de su zona (solucionando el desfase de 4h en telemetria de `EXTERIOR`).
- [index.ts](file:///c:/Dev/pristinoplant/services/scheduler/src/index.ts): Restaurar mutabilidad de `lastEmaTelemetryBatchAt` (reemplazar `const` por `let`) y actualizar timestamp al recibir lotes de telemetria climatica del EMA.
- [clean-and-align-today-telemetry.ts](file:///c:/Dev/pristinoplant/services/scheduler/src/scripts/clean-and-align-today-telemetry.ts): Crear script de sobreescritura in-situ para desduplicar y alinear la telemetria climatica en InfluxDB por ventana movil usando setFloatField.

+++ services/scheduler/src/scripts/clean-and-align-today-telemetry.ts
@@ -140,7 +140,7 @@ async function main() {
 
         if (newTemp !== null) point.setFloatField('temperature', newTemp)
         if (newHum !== null) point.setFloatField('humidity', newHum)
-        if (newLux !== null) point.setIntField('illuminance', newLux)
+        if (newLux !== null) point.setFloatField('illuminance', newLux)
         if (current.rainIntensity !== null) point.setFloatField('rain_intensity', current.rainIntensity)
 
         pointsToWrite.push(point)
